requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
cloudflare==2.8.15
idna==3.2
jsonlines==2.0.0
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
